+++ packages/graphql/src/translate/queryAST/ast/filters/RelationshipFilter.ts
@@ -246,7 +246,7 @@ export class RelationshipFilter extends Filter {
             const returnVar = new Cypher.Variable();
             returnVariables.push(returnVar);
             const nestedSubqueries = f.getSubqueries(context).map((sq) => {
-                return new Cypher.Call(sq).innerWith(context.target);
+                return new Cypher.Call(sq).importWith(context.target);
             });
 
             let predicate = f.getPredicate(context);