+++ packages/graphql/src/translate/queryAST/ast/operations/ConnectionReadOperation.ts
@@ -122,9 +122,16 @@ export class ConnectionReadOperation extends Operation {
             selectionClause = new Cypher.With("*");
         }
 
-        const authFilterSubqueries = this.getAuthFilterSubqueries(nestedContext).map((sq) =>
-            new Cypher.Call(sq).innerWith(nestedContext.target)
-        );
+        const authFilterSubqueries = this.getAuthFilterSubqueries(nestedContext).map((sq) => {
+            return new Cypher.Call(sq).importWith(nestedContext.target);
+        });
+
+        const normalFilterSubqueries = this.getFilterSubqueries(nestedContext).map((sq) => {
+            return new Cypher.Call(sq).importWith(nestedContext.target);
+        });
+
+        const filtersSubqueries = [...authFilterSubqueries, ...normalFilterSubqueries];
+      
         const edgesVar = new Cypher.NamedVariable("edges");
         const totalCount = new Cypher.NamedVariable("totalCount");
         const edgesProjectionVar = new Cypher.Variable();
@@ -137,7 +144,7 @@ export class ConnectionReadOperation extends Operation {
 
         let withWhere: Cypher.With | undefined;
 
-        if (authFilterSubqueries.length > 0) {
+        if (filtersSubqueries.length > 0) {
             withWhere = new Cypher.With("*");
             this.addFiltersToClause(withWhere, nestedContext);
         } else {
@@ -170,7 +177,7 @@ export class ConnectionReadOperation extends Operation {
                 Cypher.concat(
                     ...extraMatches,
                     selectionClause,
-                    ...authFilterSubqueries,
+                    ...filtersSubqueries,
                     withWhere,
                     withCollectEdgesAndTotalCount,
                     unwindAndProjectionSubquery,
@@ -183,6 +190,10 @@ export class ConnectionReadOperation extends Operation {
 
     protected getAuthFilterSubqueries(context: QueryASTContext): Cypher.Clause[] {
         return this.authFilters.flatMap((f) => f.getSubqueries(context));
+    }
+
+    protected getFilterSubqueries(context: QueryASTContext): Cypher.Clause[] {
+        return this.filters.flatMap((f) => f.getSubqueries(context));
     }
 
     protected getAuthFilterPredicate(context: QueryASTContext): Cypher.Predicate[] {
@@ -219,7 +230,7 @@ export class ConnectionReadOperation extends Operation {
                 ...postPaginationSubqueries,
                 new Cypher.Return([Cypher.collect(edgeProjectionMap), returnVar])
             )
-        ).innerWith(edgesVar);
+        ).importWith(edgesVar);
     }
 
     private createProjectionMapForEdge(context: QueryASTContext<Cypher.Node>): Cypher.Map {